The file 'monopoly.py' was modularized
As requested, I modularized monopoly.py. I created the invidual files:

board.py
cards.py

I did not move Player to an individual file because there was an existing
'player.py'.

Resolves issue: #1
